backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
python-dotenv==1.0.0
gunicorn==21.2.0

